refactor: separate Reducer+Lockman into dedicated files
- Extract LockmanReducer to dedicated LockmanReducer.swift (77 lines)
- Extract LockmanDynamicConditionReducer to dedicated LockmanDynamicConditionReducer.swift (467 lines)
- Keep Reducer+Lockman.swift focused on extension methods only (377 lines)
- Improve code organization and maintainability
- Reduce single file complexity from 916 lines to 3 focused files
- All tests pass and build succeeds
